Autodetect network interfaces if no taps were specified
That was my only grief with speedometer compared to tools
like iftop: Having to find out the network device name
beforehand.

This patch changes that: if no taps are specified on the
command line, go through /proc/net/dev, filter out the
interesting devices, and use those (both RX and TX for
each).